Move NestedMenu component to dedicated file
